# src/main/resources/static/swagger.yml
openapi: 3.0.4
info:
  title: itPlace API Document
  description: |-
    지도 기반 내 주변 LG유플러스 멤버십 혜택 정보 조회 서비스 'itPlace'의 API 명세서입니다.
    
    모든 API 응답은 표준 응답 포맷(ApiResponse)을 따릅니다.
  version: 1.0.0
servers:
  - url: https://userapi.itplace.click/api/v1
    description: 개발 서버

  - url: http://localhost:8080/api/v1
    description: 로컬 테스트 서버
...
paths:
...
  /auth/signUp:
    post:
      tags:
        - Auth
      summary: 최종 회원가입
      description: |
        이메일, 비밀번호, 이름 등을 사용하여 새로운 사용자를 등록합니다.
        
        이메일 인증이 사전에 완료되어야 하며, 인증 시 사용한 정보를 기반으로 가입을 처리합니다.
      operationId: signUp
      requestBody:
        description: 회원가입에 필요한 사용자 정보
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
            example:
              name: "홍길동"
              email: "user@example.com"
              phoneNumber: "01011112222"
              password: "!Password123"
              passwordConfirm: "!Password123"
              gender: "MALE"
              birthday: "20000101"
              membershipId: "1234567812345678"
      responses:
        '201':
          description: 회원가입 성공.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "SIGNUP_SUCCESS"
                status: "OK"
                message: "성공적으로 회원가입이 완료 되었습니다."
                data: null
                timestamp: "2025-07-04T14:25:38.123Z"
        '400':
          description: 비밀번호 검증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "PASSWORD_MISMATCH"
                status: "BAD_REQUEST"
                message: "비밀번호가 일치하지 않습니다."
                data: null
                timestamp: "2025-07-04T14:25:38.456Z"
        '409':
          description: 데이터 충돌 (중복된 리소스)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                emailExists:
                  summary: 이메일 중복
                  value:
                    code: "DUPLICATE_EMAIL"
                    status: "CONFLICT"
                    message: "이미 사용 중인 이메일입니다."
                    data: null
                    timestamp: "2025-07-04T14:25:38.789Z"
                phoneNumberExists:
                  summary: 전화번호 중복
                  value:
                    code: "DUPLICATE_PHONE_NUMBER"
                    status: "CONFLICT"
                    message: "이미 사용 중인 전화번호입니다."
                    data: null
                    timestamp: "2025-07-04T14:25:38.912Z"
  /auth/loadUplusData:
    post:
      tags:
        - Auth
      summary: 유플러스 데이터 가져오기
      description: 유플러스에 기존 데이터가 존재할 경우 해당 데이터를 가져올 것인지 응답을 받습니다.
      operationId: loadUplusData
      requestBody:
        description: 유플러스 데이터를 요청
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UplusDataRequest'
            example:
              name: "홍길동"
              phoneNumber: "01012345678"
      responses:
        '200':
          description: 정상 처리
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UplusDataResponse'
              example:
                code: "UPLUS_DATA_FOUND"
                status: "OK"
                message: "데이터를 찾았습니다."
                data:
                  name: "홍길동"
                  phoneNumber: "01012345678"
                  gender: "MALE"
                  birthday: "20000101"
                  membershipId: "1234567812345678"
                timestamp: "2025-07-04T11:27:06.789Z"
        '400':
          description: 비정상 처리
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "UPLUS_DATA_NOT_FOUND"
                status: "BAD_REQUEST"
                message: "데이터를 찾지 못했습니다."
                data: null
                timestamp: "2025-07-04T12:23:00.456Z"
...
  /users:
    get:
      tags:
        - User
      summary: 사용자 정보 조회
      description: 사용자 ID를 기반으로 사용자의 기본 정보와 멤버십 정보를 조회합니다.
      responses:
        '200':
          description: 사용자 정보 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserInfoResponse'
              examples:
                success:
                  summary: 사용자 정보 조회 성공 예시
                  value:
                    code: USER_INFO_SUCCESS
                    status: OK
                    message: 사용자 정보 조회에 성공했습니다.
                    data:
                      id: 1
                      name: "홍길동"
                      email: "hong@example.com"
                      phoneNumber: "01012345678"
                      gender: "MALE"
                      birthday: "1990-01-01"
                      membershipId: "M001"
                      membershipGrade: "BASIC"
                    timestamp: '2025-07-16T10:00:00.000Z'
    delete:
      tags:
        - User
      summary: 사용자 탈퇴
      description: 현재 로그인된 사용자를 탈퇴 처리합니다. 즐겨찾기, 소셜 계정 연관 정보가 모두 삭제되며 복구할 수 없습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
            example:
              password: "1234"
      responses:
        '200':
          description: 사용자 탈퇴 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: 사용자 탈퇴 성공 예시
                  value:
                    data: null
                    code: USER_WITHDRAWAL_SUCCESS
                    status: OK
                    message: 회원 탈퇴가 완료되었습니다.
                    timestamp: '2025-07-23T15:00:00.000Z'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                userNotFound:
                  summary: 사용자 없음 예시
                  value:
                    data: null
                    code: USER_NOT_FOUND
                    status: NOT_FOUND
                    message: 사용자 정보를 찾을 수 없습니다.
                    timestamp: '2025-07-23T15:00:00.000Z'
        '400':
          description: 비밀번호 불일치
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                userNotFound:
                  summary: 사용자 없음 예시
                  value:
                    data: null
                    code: PASSWORD_MISMATCH
                    status: BAD_REQUEST
                    message: 비밀번호가 일치하지 않습니다.
                    timestamp: '2025-07-23T15:00:00.000Z'
  /users/findEmail:
    post:
      tags:
        - User
      summary: 아이디(이메일) 찾기 인증코드 받기
      description: 사용자 전화번호를 인증하여 가입된 이메일을 찾기 위한 인증 코드를 발급 받습니다.
      operationId: findEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsVerificationRequest'
            example:
              name: "홍길동"
              phoneNumber: "010-1234-5678"
      responses:
        '200':
          description: 인증 코드 발송 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "SMS_SEND_SUCCESS"
                status: "OK"
                message: "문자 인증 코드가 성공적으로 발송되었습니다."
                data: null
                timestamp: "2025-07-04T11:27:06.789Z"
  /users/findEmail/confirm:
    post:
      tags:
        - User
      summary: 아이디(이메일) 찾기 인증코드 확인
      description: 사용자 전화번호로 받은 인증코드를 인증합니다.
      operationId: findEmailConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindEmailConfirmRequest'
            example:
              name: "홍길동"
              phoneNumber: "010-1234-5678"
              verificationCode: "123456"
      responses:
        '200':
          description: 인증 코드 확인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FindEmailResponse'
              example:
                code: "EMAIL_FIND_SUCCESS"
                status: "OK"
                message: "이메일 찾기를 성공했습니다."
                data:
                  email: "user@example.com"
                timestamp: "2025-07-04T11:27:06.789Z"
  /users/findPassword:
    post:
      tags:
        - User
      summary: 비밀번호 찾기 인증코드 받기
      description: 사용자 이메일로 인증 코드를 받습니다.
      operationId: findPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
            example:
              email: "user@example.com"
      responses:
        '200':
          description: 인증 코드 발송 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "EMAIL_SEND_SUCCESS"
                status: "OK"
                message: "이메일 인증 코드 발송에 성공했습니다."
                data: null
                timestamp: "2025-07-04T11:27:06.789Z"
  /users/findPasswordConfirm:
    post:
      tags:
        - User
      summary: 비밀번호 찾기 인증코드 확인
      description: 사용자 이메일로 받은 인증 코드를 인증합니다.
      operationId: findPasswordConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailConfirmRequest'
            example:
              email: "user@example.com"
              verificationCode: "123456"
      responses:
        '200':
          description: 인증 코드 확인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FindPasswordConfirmResponse'
              example:
                code: "EMAIL_VERIFICATION_SUCCESS"
                status: "OK"
                message: "이메일 인증에 성공했습니다."
                data:
                  resetPasswordToken: "42832341-acee-4bdb-8a04-b4c55f6bdccd"
                timestamp: "2025-07-04T11:27:06.789Z"
  /users/resetPassword:
    post:
      tags:
        - User
      summary: 비밀번호 재설정
      description: 인증이 끝난 사용자의 비밀번호를 재설정합니다.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            example:
              resetPasswordToken: "42832341-acee-4bdb-8a04-b4c55f6bdccd"
              email: "user@example.com"
              newPassword: "newPassword123"
              newPasswordConfirm: "newPassword123"
      responses:
        '200':
          description: 비밀번호 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: "RESET_PASSWORD_SUCCESS"
                status: "OK"
                message: "비밀번호가 변경되었습니다."
                data: null
                timestamp: "2025-07-04T11:27:06.789Z"
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: "서버에서 정의한 응답 코드"
          example: "SUCCESS_200_1"
        status:
          type: string
          description: "HTTP 상태 메시지"
          example: "OK"
        message:
          type: string
          description: "응답에 대한 설명 메시지"
          example: "요청이 정상적으로 처리되었습니다."
        data:
          type: object
          description: "실제 응답 데이터 (nullable)"
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: "응답 시간"

    # ----------------------------------------------------
    # Request DTOs
    # ----------------------------------------------------
...
    SignUpRequest:
      type: object
      required: [ name, email, password, passwordConfirm, phoneNumber, gender, birthday ]
      properties:
        name:
          type: string
          description: "이름"
          example: "홍길동"
        email:
          type: string
          format: email
          description: "아이디로 사용할 이메일 주소"
          example: "user@example.com"
        password:
          type: string
          format: password
          description: "비밀번호 (영문, 숫자, 특수문자 포함 8자 이상)"
          example: "!Password123"
          minLength: 6
        passwordConfirm:
          type: string
          format: password
          description: "비밀번호 확인"
          example: "!Password123"
          minLength: 6
        phoneNumber:
          type: string
          format: phoneNumber
          description: "전화번호"
          example: "01011112222"
        gender:
          type: string
          description: "성별"
          enum: [ MALE, FEMALE ]
          example: "MALE"
        birthday:
          type: string
          description: "생년월일"
          example: "20000101"
        membershipId:
          type: string
          description: "멤버십 번호 16자리"
          example: "1234567812345678"
    UplusDataRequest:
      type: object
      required: [ name, phoneNumber ]
      properties:
        email:
          type: string
          format: string
          example: "홍길동"
        password:
          type: string
          format: string
          example: "01012345678"
    SmsVerificationRequest:
      type: object
      description: "문자 인증 코드 발송 요청 DTO"
      required: [ name, phoneNumber ]
      properties:
        name:
          type: string
          description: "가입자 이름"
          example: "홍길동"
        phoneNumber:
          type: string
          description: "인증을 진행할 전화번호"
          example: "01012345678"
...
    FindEmailConfirmRequest:
      type: object
      description: "문자 인증 코드 발송 요청 DTO"
      required: [ name, phoneNumber, verificationCode ]
      properties:
        name:
          type: string
          description: "가입자 이름"
          example: "홍길동"
        phoneNumber:
          type: string
          description: "인증을 진행할 전화번호"
          example: "01012345678"
        verificationCode:
          type: string
          description: "인증 코드"
          example: "123456"
    EmailVerificationRequest:
      type: object
      description: "이메일 인증 코드 발송 요청 DTO"
      required: [ email ]
      properties:
        email:
          type: string
          format: email
          description: "아이디로 사용할 이메일"
          example: "user@example.com"
    EmailConfirmRequest:
      type: object
      description: "이메일 인증 코드 확인 요청 DTO"
      required: [ email, verificationCode ]
      properties:
        email:
          type: string
          format: email
          description: "아이디로 사용할 이메일"
          example: "user@example.com"
        verificationCode:
          type: string
          description: "인증 번호"
          example: "123456"
    ResetPasswordRequest:
      type: object
      description: "이메일 인증 코드 확인 요청 DTO"
      required: [ resetPasswordToken, email, newPassword, newPasswordConfirm ]
      properties:
        resetPasswordToken:
          type: string
          description: "비밀번호 재설정 토큰"
          example: "42832341-acee-4bdb-8a04-b4c55f6bdccd"
        email:
          type: string
          format: email
          description: "재설정할 이메일"
          example: "user@example.com"
        newPassword:
          type: string
          description: "새로운 비밀번호"
          example: "!newPassword123"
        newPasswordConfirm:
          type: string
          format: email
          description: "새로운 비밀번호 확인"
          example: "!newPassword123"
...
    WithdrawRequest:
      type: object
      properties:
        password:
          type: string
          description: 사용자가 입력한 비밀번호
...
    FindEmailResponse:
      type: object
      required: [ email ]
      properties:
        name:
          type: string
          format: email
          example: "user@example.com"
    FindPasswordConfirmResponse:
      type: object
      required: [ resetPasswordToken ]
      properties:
        resetPasswordToken:
          type: string
          format: string
          example: "42832341-acee-4bdb-8a04-b4c55f6bdccd"
...
    UplusDataResponse:
      type: object
      properties:
        name:
          type: string
          example: "홍길동"
        phoneNumber:
          type: string
          example: "01012345678"
        gender:
          type: string
          enum: [ MALE, FEMALE ]
          example: "MALE"
        birthday:
          type: string
          example: "20000101"
        membershipId:
          type: integer
          example: "1234567812345678"
...
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 사용자 ID
          example: 1
        name:
          type: string
          description: 사용자 이름
          example: "홍길동"
        email:
          type: string
          format: email
          description: 사용자 이메일
          example: "hong@example.com"
        phoneNumber:
          type: string
          description: 사용자 전화번호
          example: "010-1234-5678"
        gender:
          type: string
          description: 사용자 성별
          enum: [ MALE, FEMALE, OTHER ]
          example: "MALE"
        birthday:
          type: string
          format: date
          description: 사용자 생년월일
          example: "1990-01-01"
        membershipId:
          type: string
          description: 멤버십 ID (없는 경우 null)
          example: "M001"
          nullable: true
        membershipGrade:
          type: string
          description: 멤버십 등급 (없는 경우 null)
          enum: [ VVIP, VIP, BASIC ]
          example: "BASIC"
          nullable: true
...